refactored to create rastertile and vectortile classes
